+++ examples/cra-kitchen-sink/src/stories/index.stories.js
@@ -12,8 +12,8 @@ const InfoButton = () => (
       fontFamily: 'sans-serif',
       fontSize: 12,
       textDecoration: 'none',
-      background: 'rgb(34, 136, 204)',
-      color: 'rgb(255, 255, 255)',
+      background: '#027ac5',
+      color: '#fff',
       padding: '5px 15px',
       margin: 10,
       borderRadius: '0px 0px 0px 5px',
@@ -36,17 +36,20 @@ storiesOf('Button', module)
   })
   .add(
     'with new info',
-    withInfo(
-      'Use the [info addon](https://github.com/storybooks/storybook/tree/master/addons/info) with its new painless API.'
-    )(context => (
+    context => (
       <Container>
         <span>
           click the <InfoButton /> label in top right for info about "{context.name}"
         </span>
       </Container>
-    )),
+    ),
     {
       notes: 'Composition: Info(Notes())',
       options: { selectedPanel: 'storybook/info/panel' },
+      decorators: [
+        withInfo(
+          'Use the [info addon](https://github.com/storybooks/storybook/tree/master/addons/info) with its new painless API.'
+        ),
+      ],
     }
   );