+++ lib/api/src/store.ts
@@ -1,20 +1,24 @@
-import { localStorage, sessionStorage } from 'global';
-import { parse, stringify } from 'telejson';
+import store, { StoreAPI } from 'store2';
+import storeSetup from './lib/store-setup';
+
+// setting up the store, overriding set and get to use telejson
+// @ts-ignore
+storeSetup(store._);
 
 export const STORAGE_KEY = '@storybook/ui/store';
 
 import { State } from './index';
 
-function get(storage: Storage) {
-  const serialized = storage.getItem(STORAGE_KEY);
-  return serialized ? parse(serialized) : {};
+function get(storage: StoreAPI) {
+  const data = storage.get(STORAGE_KEY);
+  return data || {};
 }
 
-function set(storage: Storage, value: Patch) {
-  storage.setItem(STORAGE_KEY, stringify(value, { maxDepth: 50 }));
+function set(storage: StoreAPI, value: Patch) {
+  storage.set(STORAGE_KEY, value);
 }
 
-function update(storage: Storage, patch: Patch) {
+function update(storage: StoreAPI, patch: Patch) {
   const previous = get(storage);
   // Apply the same behaviour as react here
   set(storage, { ...previous, ...patch });
@@ -22,18 +26,6 @@ function update(storage: Storage, patch: Patch) {
 
 type GetState = () => State;
 type SetState = (a: any, b: any) => any;
-
-interface Storage {
-  getItem(key: string): string | undefined;
-  setItem(
-    key: string,
-    value: string
-  ):
-    | {
-        [key: string]: any;
-      }
-    | undefined;
-}
 
 interface Upstream {
   getState: GetState;
@@ -68,7 +60,7 @@ export default class Store {
     // We don't only merge at the very top level (the same way as React setState)
     // when you set keys, so it makes sense to do the same in combining the two storage modes
     // Really, you shouldn't store the same key in both places
-    return { ...base, ...get(localStorage), ...get(sessionStorage) };
+    return { ...base, ...get(store.local), ...get(store.session) };
   }
 
   getState() {
@@ -112,7 +104,7 @@ export default class Store {
     });
 
     if (persistence !== 'none') {
-      const storage = persistence === 'session' ? sessionStorage : localStorage;
+      const storage = persistence === 'session' ? store.session : store.local;
       await update(storage, delta);
     }